Allow dependency versions to be more flexible.
hdf5 updates slowly, and because that depends on ndarray, the dependent
ndarray version might be old. I had tried to make hyperbeam use only one
version of ndarray in the past with a greater-than and less-than
restraint, but this issue causes pain elsewhere (specifically, other
crates that depend on hyperbeam might depend on a newer version of
ndarray). My strategy in this commit is to effectively ignore ndarray
coming out of hdf5.

To help with potential issues when converting from hdf5 vectors to
ndarrays, I've added some runtime checks, but these aren't really
robust. As it always has, hyperbeam continues to assume a lot about the
structure of the incoming HDF5 file.