Add control over dipole elements and bump version.
Previous versions of hyperbeam only allowed an MWA dipole to be
"on" (usually with amp of 1) or off (amp of 0). However, a dipole has
two elements (X and Y); with this commit, those elements can now be
controlled. If 32 amps are supplied, the first 16 are for the X dipole
elements, the next 16 the Y dipole elements.

The existing interface to hyperbeam has not changed. Rust users can
simply pass 32 amps instead of 16 to get the full control, but C/Python
users must use additional functions.

With this additional functionality, I made the types a little tighter
through FEE beam code (specifically, amps is now always a 32-element
array), which helps to safeguard against using the wrong amps somewhere.
With my improved Rust knowledge, I was also able to remove a couple of
allocations.

Add control over dipole elements and bump version.

Previous versions of hyperbeam only allowed an MWA dipole to be
"on" (usually with amp of 1) or off (amp of 0). However, a dipole has
two elements (X and Y); with this commit, those elements can now be
controlled. If 32 amps are supplied, the first 16 are for the X dipole
elements, the next 16 the Y dipole elements.

The existing interface to hyperbeam has not changed. Rust users can
simply pass 32 amps instead of 16 to get the full control, but C/Python
users must use additional functions.

With this additional functionality, I made the types a little tighter
through FEE beam code (specifically, amps is now always a 32-element
array), which helps to safeguard against using the wrong amps somewhere.
With my improved Rust knowledge, I was also able to remove a couple of
allocations.