Add CPU code for the analytic beam.
mwa_pb and the RTS disagree on how the beam is implemented, so both
versions are available. You're welcome.

A future release should rename the FEE FFI functions to be unambiguously to do
with FEE (e.g. calc_jones -> fee_calc_jones). This is not done here so we don't
break semver.

The `mut` requirement on `self` for python functions has been
removed. Similarly, the FEE FFI beam is no longer accessed as mutable. I'm
pretty sure this doesn't change anything.